requirements.txt
asgiref==3.9.1
aspose-slides==25.7.0
aspose-words==25.8.0
blinker==1.7.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==39.0.0
dbus-python==1.3.2
defusedxml==0.7.1
diff-match-patch==20200713
distro==1.9.0
distro-info==1.7+build1
Django==4.1.10
django-cors-headers==3.13.0
django-dotenv==1.4.2
django-import-export==3.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
et-xmlfile==1.1.0
graphene==3.4.3
graphene-django==3.2.3
graphql-core==3.2.6
graphql-relay==3.2.0
grpcio==1.76.0
grpcio-tools==1.76.0
httplib2==0.20.4
idna==3.4
launchpadlib==1.11.0
lazr.restfulclient==0.14.6
lazr.uri==1.0.6
lxml==4.9.2
MarkupPy==1.14
numpy==1.26.4
oauthlib==3.2.2
odfpy==1.4.1
opencv-python==4.6.0.66
openpyxl==3.0.10
packaging==22.0
pandas==1.5.2
Pillow==10.0.1
promise==2.3
protobuf==6.33.0
psycopg2-binary==2.9.10
pycparser==2.21
PyGObject==3.48.2
PyJWT==2.6.0
pyparsing==3.1.1
pypdf==3.8.1
pypng==0.20220715.0
python-apt==2.7.7+ubuntu5
python-dateutil==2.8.2
python-decouple==3.7
python-pptx==0.6.21
python-pptx-text-replacer==0.0.6
pytz==2022.6
PyYAML==6.0
qrcode==7.4.2
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
tablib==3.2.1
text-unidecode==1.3
tqdm==4.64.1
twilio==7.17.0
typing_extensions==4.15.0
unattended-upgrades==0.1
urllib3==1.26.13
wadllib==1.3.6
xlrd==2.0.1
XlsxWriter==3.0.8
xlwt==1.3.0
